Move o zombificator para o namespace untied e refatora o act_as_zombie para acts_as_zombie
