+++ app/models/user.rb
@@ -1,9 +1,9 @@
 class User < ActiveRecord::Base
-  include Zombificator::ActAsZombie
+  include Untied::Zombificator::ActAsZombie
 
   attr_reader :thumbnail_remote_url
 
-  act_as_zombie
+  acts_as_zombie
 
   attr_accessible :core_id, :login, :email, :first_name, :last_name, :role,
   :thumbnail, :client_applications
